+++ activerecord/lib/active_record/connection_adapters/postgresql_adapter.rb
@@ -302,6 +302,10 @@ module ActiveRecord
         true
       end
 
+      def supports_partial_index?
+        true
+      end
+
       class StatementPool < ConnectionAdapters::StatementPool
         def initialize(connection, max)
           super